Add support for defaultApp
