+++ generators/app/index.js
@@ -23,7 +23,6 @@ const jsonfile = require('jsonfile');
 const semver = require('semver');
 const shelljs = require('shelljs');
 const BaseGenerator = require('generator-jhipster/generators/generator-base');
-const prompts = require('./prompts');
 const modifyPackage = require('modify-package-dependencies');
 const spawn = require('cross-spawn');
 const fs = require('fs');
@@ -54,6 +53,12 @@ module.exports = class extends BaseGenerator {
 
     get initializing() {
         return {
+            init(args) {
+                if (args === 'default') {
+                    this.defaultApp = true;
+                    this.interactive = false;
+                }
+            },
             readConfig() {
                 this.jhipsterAppConfig = this.getJhipsterAppConfig();
             },
@@ -68,11 +73,46 @@ module.exports = class extends BaseGenerator {
         };
     }
 
-    get prompting() {
-        return {
-            askForProjectName: prompts.askForAppName,
-            askForPath: prompts.askForPath
-        };
+    prompting() {
+        const done = this.async();
+        const messageAskForPath = 'Enter the directory where your JHipster app is located:';
+        const prompts = [
+            {
+                type: 'input',
+                name: 'appName',
+                message: 'What do you want to name your Ionic application?',
+                default: 'ionic4j'
+            },
+            {
+                type: 'input',
+                name: 'directoryPath',
+                message: messageAskForPath,
+                default: 'backend',
+                validate: (input) => {
+                    const path = this.destinationPath(input);
+                    if (shelljs.test('-d', path)) {
+                        const appsFolders = this.getAppFolder.call(this, input);
+                        if (appsFolders.length === 0) {
+                            return `No application found in ${path}`;
+                        }
+                        return true;
+                    }
+                    return `${path} is not a directory or doesn't exist`;
+                }
+            }];
+
+        if (this.defaultApp) {
+            this.ionicAppName = 'ionic4j';
+            this.directoryPath = 'backend';
+            done();
+        } else {
+            this.prompt(prompts).then((props) => {
+                this.ionicAppName = props.appName;
+                this.directoryPath = props.directoryPath;
+                this.appFolder = this.getAppFolder.call(this, this.directoryPath);
+                done();
+            });
+        }
     }
 
     writing() {
@@ -101,7 +141,7 @@ module.exports = class extends BaseGenerator {
             params.push('--no-deps');
             params.push('--no-git');
         }
-        spawn.sync('ionic', params, { stdio: 'inherit' });
+        spawn.sync('ionic', params, {stdio: 'inherit'});
     }
 
     install() {
@@ -235,5 +275,29 @@ module.exports = class extends BaseGenerator {
         portWarning += chalk.yellow(`    ${this.directoryPath}/webpack/webpack.dev.js\n`);
         portWarning += chalk.yellow(`    ${this.ionicAppName}/src/providers/api/api.ts\n`);
         this.log(portWarning);
+    }
+
+    /**
+     * Get App Folders
+     * @param input path to join to destination path
+     * @returns {Array} array of string representing app folders
+     */
+    getAppFolder(input) {
+        const destinationPath = this.destinationPath(input);
+        const appsFolders = [];
+
+        if (shelljs.test('-f', `${destinationPath}/.yo-rc.json`)) {
+            try {
+                const fileData = this.fs.readJSON(`${destinationPath}/.yo-rc.json`);
+                if (fileData['generator-jhipster'].baseName !== undefined) {
+                    appsFolders.push(destinationPath);
+                }
+            } catch (err) {
+                this.log(chalk.red(`The .yo-rc.json in ${destinationPath} can't be read!`));
+                this.debug('Error:', err);
+            }
+        }
+
+        return appsFolders;
     }
 };
